server/requirements.txt
flake8
autopep8
django
django-rest-framework
django-rest-auth
django-allauth
django-extensions
ipython
django-extra-fields
celery == 4.4.0
redis == 3.2.1
flower == 0.9.3
django-cors-headers == 3.1.0
pytest == 4.3.0
tornado == 4.2.0
pillow == 6.1.0
psycopg2
django-simple-history==3.0.0
djangorestframework-simplejwt
# django-rest-knox